+++ jabgui/src/main/java/org/jabref/gui/preferences/websearch/WebSearchTab.java
@@ -3,6 +3,7 @@ package org.jabref.gui.preferences.websearch;
 import java.util.List;
 import java.util.Optional;
 
+import javafx.beans.binding.Bindings;
 import javafx.beans.InvalidationListener;
 import javafx.beans.property.ReadOnlyBooleanProperty;
 import javafx.fxml.FXML;
@@ -93,6 +94,10 @@ public class WebSearchTab extends AbstractPreferenceTabView<WebSearchTabViewMode
         searchEngineUrlTemplate.setEditable(true);
 
         searchEngineTable.setItems(viewModel.getSearchEngines());
+        searchEngineTable.prefHeightProperty().bind(
+                Bindings.size(viewModel.getSearchEngines()).
+                        multiply(searchEngineTable.getFixedCellSize()).
+                        add(searchEngineTable.getFixedCellSize()));
 
         enableWebSearch.selectedProperty().bindBidirectional(viewModel.enableWebSearchProperty());
         warnAboutDuplicatesOnImport.selectedProperty().bindBidirectional(viewModel.warnAboutDuplicatesOnImportProperty());